+++ colorcoder/Testcode.java
@@ -1,0 +1,22 @@
+package colorcoder;
+
+public class Testcode {
+	
+
+    static void testNumberToPair(int pairNumber, MajorColor expectedMajor, MinorColor expectedMinor)
+    {
+        ColorPair colorPair = ColorService.GetColorFromPairNumber(pairNumber);
+        System.out.println("Got pair " + colorPair.ToString());
+        assert(colorPair.getMajor() == expectedMajor);
+        assert(colorPair.getMinor() == expectedMinor);
+    }
+
+    static void testPairToNumber(MajorColor major, MinorColor minor, int expectedPairNumber)
+    {
+        int pairNumber = ColorService.GetPairNumberFromColor(major, minor);
+        System.out.println("Got pair number " + pairNumber);
+        assert(pairNumber == expectedPairNumber);
+    }
+
+    
+}
